Move Shoryuken middleware to integrations
Move the ShoryukenMiddleware to the integrations directory, matching how
we organize other integrations: a separate hook and integrations file.
